feat(core/internal): migrate to Japa as test runner.
tap is a bit of a mess right now and its time to cut losses
