+++ .c8rc.json
@@ -1,0 +1,8 @@
+{
+    "src": "./src",
+    "all": true,
+    "reporter": [
+        "text",
+        "lcovonly"
+    ]
+}
