+++ .gitignore
@@ -4,4 +4,6 @@ node_modules/
 coverage/
 test.sh
 test.sh.sha256
-.tap/
+.tsimp/
+**/*.js
+
