+++ biome.json
@@ -1,7 +1,7 @@
 {
 	"$schema": "https://biomejs.dev/schemas/1.7.1/schema.json",
 	"files": {
-		"include": ["./src/**/*.ts", "./test/**/*.ts"],
+		"include": ["./src/**/*.ts", "bin/test.ts", "./test/**/*.ts"],
 		"ignore": ["./test/fixtures/**/*"]
 	},
 	"organizeImports": {
