+++ package.json
@@ -1,70 +1,62 @@
 {
-    "name": "codeclimate-action",
-    "version": "6.0.0",
-    "private": true,
-    "description": "Publish code coverage to Code Climate",
-    "type": "module",
-    "main": "lib/main.js",
-    "packageManager": "pnpm@9.1.3",
-    "scripts": {
-        "build": "tsc",
-        "test": "cross-env NODE_OPTIONS=\"--no-warnings\" tap run --disable-coverage",
-        "test:only": "pnpm run test --only",
-        "coverage": "pnpm run test --no-disable-coverage --allow-incomplete-coverage --coverage-report=lcovonly",
-        "coverage:html": "pnpm run coverage --coverage-report=html",
-        "biome:ci": "biome ci ./",
-        "format:check": "biome format ./",
-        "format": "biome format --write ./",
-        "lint:check": "biome lint ./",
-        "lint": "biome lint --apply ./",
-        "lint:unsafe": "biome lint --apply-unsafe ./"
-    },
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/paambaati/codeclimate-action.git"
-    },
-    "keywords": [
-        "actions",
-        "codeclimate",
-        "quality",
-        "coverage"
-    ],
-    "author": "GP <me@httgp.com>",
-    "license": "MIT",
-    "dependencies": {
-        "@actions/core": "1.10.1",
-        "@actions/exec": "1.1.1",
-        "@actions/github": "6.0.0",
-        "@actions/glob": "0.4.0",
-        "hook-std": "3.0.0",
-        "node-fetch": "3.3.2",
-        "openpgp": "5.11.1"
-    },
-    "devDependencies": {
-        "@biomejs/biome": "1.7.3",
-        "@openpgp/web-stream-tools": "0.1.1",
-        "@tapjs/core": "2.0.1",
-        "@tapjs/test": "2.0.1",
-        "@tapjs/tsx": "1.1.25",
-        "@tsconfig/node20": "20.1.4",
-        "@tsconfig/strictest": "2.0.5",
-        "@types/node": "20.12.12",
-        "@types/node-fetch": "2.6.11",
-        "@types/sinon": "17.0.3",
-        "@types/which": "3.0.3",
-        "cross-env": "7.0.3",
-        "nock": "13.5.4",
-        "sinon": "18.0.0",
-        "tap": "19.0.2",
-        "to-readable-stream": "4.0.0",
-        "ts-node": "10.9.2",
-        "typescript": "5.4.5",
-        "which": "4.0.0"
-    },
-    "tap": {
-        "plugin": [
-            "!@tapjs/typescript",
-            "@tapjs/tsx"
-        ]
-    }
+  "name": "codeclimate-action",
+  "version": "6.0.0",
+  "private": true,
+  "description": "Publish code coverage to Code Climate",
+  "type": "module",
+  "main": "lib/main.js",
+  "packageManager": "pnpm@9.1.3",
+  "scripts": {
+    "build": "tsc -p tsconfig.build.json",
+    "test": "node --import=tsimp/import --enable-source-maps bin/test.ts",
+    "coverage": "c8 pnpm run test",
+    "coverage:html": "c8 --reporter=html pnpm run test",
+    "biome:ci": "biome ci ./",
+    "format:check": "biome format ./",
+    "format": "biome format --write ./",
+    "lint:check": "biome lint ./",
+    "lint": "biome lint --apply ./",
+    "lint:unsafe": "biome lint --apply-unsafe ./"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/paambaati/codeclimate-action.git"
+  },
+  "keywords": [
+    "actions",
+    "codeclimate",
+    "quality",
+    "coverage"
+  ],
+  "author": "GP <me@httgp.com>",
+  "license": "MIT",
+  "dependencies": {
+    "@actions/core": "1.10.1",
+    "@actions/exec": "1.1.1",
+    "@actions/github": "6.0.0",
+    "@actions/glob": "0.4.0",
+    "hook-std": "3.0.0",
+    "node-fetch": "3.3.2",
+    "openpgp": "5.11.1"
+  },
+  "devDependencies": {
+    "@biomejs/biome": "1.7.3",
+    "@japa/assert": "3.0.0",
+    "@japa/runner": "3.1.4",
+    "@openpgp/web-stream-tools": "0.1.1",
+    "@tsconfig/node20": "20.1.4",
+    "@tsconfig/strictest": "2.0.5",
+    "@types/node": "20.12.13",
+    "@types/node-fetch": "2.6.11",
+    "@types/sinon": "17.0.3",
+    "@types/which": "3.0.3",
+    "c8": "9.1.0",
+    "cross-env": "7.0.3",
+    "into-stream": "8.0.1",
+    "nock": "13.5.4",
+    "sinon": "18.0.0",
+    "tsimp": "2.0.11",
+    "typescript": "5.4.5",
+    "which": "4.0.0"
+  }
 }
