+++ tsconfig.build.json
@@ -1,0 +1,12 @@
+{
+    "$schema": "https://json.schemastore.org/tsconfig",
+    "extends": [
+        "./tsconfig.json"
+    ],
+    "compilerOptions": {
+        "rootDir": "./src"
+    },
+    "include": [
+        "./src"
+    ]
+}
