+++ tsconfig.json
@@ -5,13 +5,15 @@
         "@tsconfig/strictest/tsconfig.json"
     ],
     "compilerOptions": {
-        "outDir": "./lib", /* Redirect output structure to the directory. */
-        "rootDir": "./src", /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */
-        "rootDirs": ["./src", "./test"] /* List of root folders whose combined content represents the structure of the project at runtime. */
+        "outDir": "./lib",
+        "rootDirs": [
+            "./bin",
+            "./src",
+            "./test"
+        ],
     },
     "exclude": [
-        "node_modules",
-        "**/*.test.ts",
-        "coverage/"
+        "./coverage",
+        "./lib"
     ]
 }
