--- tsconfig.test.json
@@ -1,7 +1,0 @@
-{
-    "$schema": "https://json.schemastore.org/tsconfig",
-    "extends": "./tsconfig.json",
-    "include": [
-        "./test/*.test.ts"
-    ]
-}
